moved _add_to_list method in the Combinator class
